+++ vscode-extension/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to the "drizzle-orm" extension will be documented in this fi
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [0.7.0]
+
+- Add support for MySQL
+- Fix code lens showing up in non-drizzle-kit config files
+
 ## [0.6.0]
 
 - Add Drizzle Studio command